+++ src/components/SumLine/SumLine.types.ts
@@ -1,3 +1,4 @@
+import type { StyleProps } from "@types";
 import type { Enum, Nullable } from "@ubloimmo/front-util";
 
 export type SumLineSize = "m" | "l";
@@ -57,3 +58,5 @@ export type SumLineProps = {
 };
 
 export type SumLineDefaultProps = Required<SumLineProps>;
+
+export type SumlineStyleProps = StyleProps<Pick<SumLineDefaultProps, "size">>;